+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Unreleased]
 
+ - Added the ability to sign in with apple.
  - Added more info to room creation.
 
 ## [1.15.0] - 2021-01-22